refactor : 전역-types-위치-변경
