+++ src/features/server/model/types.ts
@@ -1,5 +1,5 @@
 import type { LoginStatus } from '@/shared/model/types';
-import type { PageInfo } from '@/shared/types/PageInfo';
+import type { PageInfo } from '@/shared/model/types/PageInfo';
 
 export interface Servers {
   name: string;
