+++ src/shared/icons/EditIcon.tsx
@@ -1,4 +1,4 @@
-import type IconTypes from '../types/IconType';
+import type IconTypes from '../model/types/IconType';
 
 export const EditIcon = ({ size = 32, color = '#C7C8CE' }: IconTypes) => {
   return (
