+++ src/shared/icons/FriendsIcon.tsx
@@ -1,4 +1,4 @@
-import type IconTypes from '../types/IconType';
+import type IconTypes from '../model/types/IconType';
 
 export const FriendsIcon = ({ size = 24, color = '#9597A3' }: IconTypes) => {
   return (
