+++ src/shared/icons/PinnedMessageIcon.tsx
@@ -1,4 +1,4 @@
-import type IconTypes from '../types/IconType';
+import type IconTypes from '../model/types/IconType';
 
 export const PinnedMessageIcon = ({ size = 20, color = '#828391' }: IconTypes) => {
   return (
