+++ src/shared/icons/WasteBasketIcon.tsx
@@ -1,4 +1,4 @@
-import type IconTypes from '../types/IconType';
+import type IconTypes from '../model/types/IconType';
 
 export const WasteBasketIcon = ({ size = 32, color = '#C7C8CE' }: IconTypes) => {
   return (
