+++ src/shared/model/types/IconType.ts
