+++ src/shared/model/types/PageInfo.ts
