+++ src/shared/model/types/activity-detector.d.ts
