+++ src/shared/model/types/apiResponse.ts
